Nyquist modification for synthetic radar obs
In the kodiak version of create_obs_radar_sequence, the Nyquist observation for
a radial velocity observation was set to missing_r8, which is negative. But,
obs_def_radar_mod does not accept negative values for the Nyquist velocity.

create_obs_radar_sequence was modified to set the Nyquist velocity for synthetic
radar observations to -missing_r8, which is a large positive value (888888).